// packages/oasis-actions/src/operations/aaveV3/addresses.ts
export interface AAVEV3StrategyAddresses {
  DAI: string
  ETH: string
  WETH: string
  STETH: string
  USDC: string
  WBTC: string
  operationExecutor: string
  chainlinkEthUsdPriceFeed: string
  aavePriceOracle: string
  aavePool: string
  aaveProtocolDataProvider: string
  aaveLendingPool: string
}